Used a chore to run task at startup instead of a button
