+++ omod/src/main/webapp/pages/orders/labOrdersManifestHome.gsp
@@ -34,7 +34,6 @@
 
     def configuration = [
             [label: "Settings", iconProvider: "kenyaui", icon: "", label: "Settings", href: ui.pageLink("kenyaemrorderentry", "orders/settings")],
-            [label: "Maintenance", iconProvider: "kenyaui", icon: "", label: "Maintenance", href: ui.pageLink("kenyaemrorderentry", "orders/maintenance")],
     ]
 %>
